Create hook for sign in up navigate
