+++ packages/twenty-front/src/modules/auth/sign-in-up/hooks/useSignInUp.tsx
@@ -1,10 +1,9 @@
 import { useCallback, useState } from 'react';
 import { SubmitHandler, UseFormReturn } from 'react-hook-form';
-import { useNavigate, useParams } from 'react-router-dom';
-import { useRecoilValue } from 'recoil';
+import { useParams } from 'react-router-dom';
 
+import { useNavigateAfterSignInUp } from '@/auth/sign-in-up/hooks/useNavigateAfterSignInUp.ts';
 import { Form } from '@/auth/sign-in-up/hooks/useSignInUpForm.ts';
-import { billingState } from '@/client-config/states/billingState';
 import { AppPath } from '@/types/AppPath';
 import { PageHotkeyScope } from '@/types/PageHotkeyScope';
 import { useSnackBar } from '@/ui/feedback/snack-bar-manager/hooks/useSnackBar';
@@ -26,15 +25,13 @@ export enum SignInUpStep {
 }
 
 export const useSignInUp = (form: UseFormReturn<Form>) => {
-  const navigate = useNavigate();
-
   const { enqueueSnackBar } = useSnackBar();
 
   const isMatchingLocation = useIsMatchingLocation();
 
-  const billing = useRecoilValue(billingState);
-
   const workspaceInviteHash = useParams().workspaceInviteHash;
+
+  const { navigateAfterSignInUp } = useNavigateAfterSignInUp();
 
   const [signInUpStep, setSignInUpStep] = useState<SignInUpStep>(
     SignInUpStep.Init,
@@ -103,20 +100,7 @@ export const useSignInUp = (form: UseFormReturn<Form>) => {
                 workspaceInviteHash,
               );
 
-        if (
-          billing?.isBillingEnabled &&
-          currentWorkspace.subscriptionStatus !== 'active'
-        ) {
-          navigate(AppPath.PlanRequired);
-          return;
-        }
-
-        if (currentWorkspace.activationStatus === 'active') {
-          navigate(AppPath.Index);
-          return;
-        }
-
-        navigate(AppPath.CreateWorkspace);
+        navigateAfterSignInUp(currentWorkspace);
       } catch (err: any) {
         enqueueSnackBar(err?.message, {
           variant: 'error',
@@ -128,8 +112,7 @@ export const useSignInUp = (form: UseFormReturn<Form>) => {
       signInWithCredentials,
       signUpWithCredentials,
       workspaceInviteHash,
-      billing?.isBillingEnabled,
-      navigate,
+      navigateAfterSignInUp,
       enqueueSnackBar,
     ],
   );
